+++ R/zzz.R
@@ -7,6 +7,11 @@
 #' @importFrom stats rnorm
 #' @importFrom utils installed.packages tail
 #'
+#'
+#' @section Documentation:
+#' Start by reading the [Large-Scale Benchmarking chapter](https://mlr3book.mlr-org.com/chapters/chapter11/large-scale_benchmarking.html)
+#' from the mlr3book.
+#'
 #' @section mlr3 Integration:
 #' This package adds the [mlr3::Task] `"oml"` and the [mlr3::Resampling] `"oml"` to
 #' [mlr3::mlr_tasks] and [mlr3::mlr_resamplings], respectively.