+++ tests/testthat/test_OMLObject.R
@@ -1,14 +1,17 @@
 skip_on_cran()
 
-test_that("Caching is disabled when cache directory does not exist (anymore)", {
+test_that("Caching directory is correctly handled when it does not exist", {
   old_threshold = lg$threshold
   lg$set_threshold("info")
   on.exit({lg$set_threshold(old_threshold)}, add = TRUE)
   cache_dir = tempfile()
-  odata = odt(31, cache = cache_dir)
+  odata = with_cache(odt(31), cache = cache_dir)
   capture.output(odata$desc)
   unlink(cache_dir, recursive = TRUE)
-  out = capture.output(odata$qualities)
-  expect_true(grepl(out[[1L]], pattern = "disabling caching for this object"))
-  expect_false(odata$cache_dir)
+  dir1 = tempfile()
+  out = capture.output(with_cache(odata$qualities, cache = dir1))
+  expect_true(grepl(out[[1L]], pattern = "changed since initializing this object"))
+  expect_true(grepl(out[[1L]], pattern = dir1))
+  expect_true(grepl(out[[1L]], pattern = cache_dir))
+  expect_equal(odata$cache_dir, dir1)
 })